[7015] Implement support -1 in item_template.stackable and item_template.maxcount fields. By VladimirMangos.
--HG--
branch : trunk

+++ src/game/Guild.h
@@ -220,12 +220,12 @@ struct GuildBankTab
 
 struct GuildItemPosCount
 {
-    GuildItemPosCount(uint8 _slot, uint8 _count) : slot(_slot), count(_count) {}
+    GuildItemPosCount(uint8 _slot, uint32 _count) : slot(_slot), count(_count) {}
 
     bool isContainedIn(std::vector<GuildItemPosCount> const& vec) const;
 
     uint8 slot;
-    uint8 count;
+    uint32 count;
 };
 typedef std::vector<GuildItemPosCount> GuildItemPosCountVec;
